Memory Efficient Stable String Gather
A program that takes in a string and an index. The string can contain both numbers, symbols and letters and will return the same string with the numbers at the index in the string with all the characters being in the same order as they were in the string. Ex. 3 12abc34def would return abc1234def. This program was made to focus on memory conservation since it does not make any copies of the string, however, this is at the cost of speed as it has to loop through multiple times.